+++ examples/fallen showcase/main.js
@@ -5,21 +5,24 @@ import { TileWin } from "/ruitk/tileWin/tileWin.js";
 let tileWin = new TileWin();
 let elements = new Elements();
 
-
-
 // styles
 import { FallenBase } from "/ruitk/themes/fallen/base.js";
 import { FallenTileWin } from "/ruitk/themes/fallen/tileWin.js";
+import { FallenInput } from "/ruitk/themes/fallen/input.js";
 
-FallenBase.init();
+
 elements.addElements(FallenBase.getElements());
 elements.addElements(FallenTileWin.getElements());
+elements.addElements(FallenInput.getElements());
 
 
+//tileWin.updateConfig({
+//    tileRowType : ["scroll", "scroll", "fixed"],
+//});
 
 elements.append("body", elements.makeElements(`<backgroundImg>{}`));
 
-tileWin.createTile("TextWindow", 1, 0, 1, 0, elements.makeElements(`
+tileWin.createTile("TextWindow", 1, 1, elements.makeElements(`
     <tile>{"content" : 
         <grid>{"c" : "auto auto", "r" : "auto auto auto auto auto auto" , "content" : 
             <h1>{"content" : "h1 h1 h1 h1 h1 h1 h1", "c" : "1", "r" : "1"}
@@ -30,15 +33,25 @@ tileWin.createTile("TextWindow", 1, 0, 1, 0, elements.makeElements(`
             <p2>{"content" : "p2 p2 p2 p2 p2 p2 p2", "c" : "1", "r" : "5"}
             <p3>{"content" : "p3 p3 p3 p3 p3 p3 p3", "c" : "1", "r" : "6"}
             
-            <h1>{"content" : <b>{"content" : "Bold"} <i>{"content" : " italic"}, "textAlign" : "right", "c" : "2", "r" : "1"}
-            <h2>{"content" : <b>{"content" : "Bold"} <i>{"content" : " italic"}, "textAlign" : "right", "c" : "2", "r" : "2"}
-            <h3>{"content" : <b>{"content" : "Bold"} <i>{"content" : " italic"}, "textAlign" : "right", "c" : "2", "r" : "3"}
+            <h1>{"content" : <b>{"content" : "Bold"} <u>{"content" : " UnderLine"} <i>{"content" : " italic "} <a>{"content" : "Link", "href" : "https://www.example.com"}, "textAlign" : "right", "c" : "2", "r" : "1"}
+            <h2>{"content" : <b>{"content" : "Bold"} <u>{"content" : " UnderLine"} <i>{"content" : " italic "} <a>{"content" : "Link", "href" : "https://www.example.com"}, "textAlign" : "right", "c" : "2", "r" : "2"}
+            <h3>{"content" : <b>{"content" : "Bold"} <u>{"content" : " UnderLine"} <i>{"content" : " italic "} <a>{"content" : "Link", "href" : "https://www.example.com"}, "textAlign" : "right", "c" : "2", "r" : "3"}
 
-            <p1>{"content" : <b>{"content" : "Bold"} <i>{"content" : " italic"}, "textAlign" : "right", "c" : "2", "r" : "4"}
-            <p2>{"content" : <b>{"content" : "Bold"} <i>{"content" : " italic"}, "textAlign" : "right", "c" : "2", "r" : "5"}
-            <p3>{"content" : <b>{"content" : "Bold"} <i>{"content" : " italic"}, "textAlign" : "right", "c" : "2", "r" : "6"}
+            <p1>{"content" : <b>{"content" : "Bold"} <u>{"content" : " UnderLine"} <i>{"content" : " italic "} <a>{"content" : "Link", "href" : "https://www.example.com"}, "textAlign" : "right", "c" : "2", "r" : "4"}
+            <p2>{"content" : <b>{"content" : "Bold"} <u>{"content" : " UnderLine"} <i>{"content" : " italic "} <a>{"content" : "Link", "href" : "https://www.example.com"}, "textAlign" : "right", "c" : "2", "r" : "5"}
+            <p3>{"content" : <b>{"content" : "Bold"} <u>{"content" : " UnderLine"} <i>{"content" : " italic "} <a>{"content" : "Link", "href" : "https://www.example.com"}, "textAlign" : "right", "c" : "2", "r" : "6"}
         }
     }   
-    `));
-tileWin.createTile("Tile the 2nd", 2, 0, 1, 0, elements.makeElements(`<tile>{}`));
+`));
+
+tileWin.createTile("InputTest", 1, 2, elements.makeElements(`
+    <tile>{"content" : 
+        <h1>{"content" : "Sign in", "textAlign" : "center"}
+        <textbox>{"f" : "signInForm", "name" : "email", "placeholder" : "enter Your Email", "title" : "Email"}
+        <textbox>{"f" : "signInForm", "name" : "password", "placeholder" : "enter Your Password", "title" : "Password"}
+        <submit>{"f" : "signInForm", "content" : "Sign In", "onClick", "callbackFunction that I'm to lazy to write"}
+    }
+`));
+
+tileWin.createTile("Tile the 2nd", 2, 1, elements.makeElements(`<tile>{}`));
 tileWin.update();