+++ doc/changelog.md
@@ -1,3 +1,29 @@
+# 0.7
+#### Licence Changes
+ - Changed licence from Apache License 2.0 to the GPL V3
+#### tileWin.js Changes
+ - Fixed an issue with hover not working in scroll tiles in `updateTest` 
+ - Changed `updateTest` to use the old fixed tile rendering method
+ - Replace `update` with `updateTest` as it is deemed stable enough 
+ - Removed `xNudge` and `yNudge` from `createTile`
+ - Removed `tileNudgeFirst` and `tileNudgeSwap` from `configStore`
+ - Added a function called `destory` to destroy a tile
+ - Added a function called `destoryAll` to destroy all tiles
+#### Fallen Changes
+ - Fixed an issue with the `tile` element overflowing when using `updateTest`
+ - Added an element to base called `u` for underline
+ - Added an element to base called `a` for links
+ - Changed wallpaper to a picture done by a friend of mine
+ - Removed the `text` property from the base element
+ - Removed the `init` function in `base.js`
+ - Renamed all colors in fallen
+#### style.js changes
+ - Added a new flag called `jsHover`
+#### tile.js Changes
+ - Better error handling in the `append` function
+#### documentation Changes
+ - Made a full rundown on fallen at [Fallen Summary](fallen/Fallen%20Summary.md)
+
 ## 0.6.3
 #### Fallen Changes
  - Changed all elements to use `elementCount` in there ID
@@ -52,8 +78,8 @@
 #### style.js changes
  - Fixed a issue with the `declare` not adding new lines afterward
 #### documentation Changes 
- - Added [1. Fallen Summary](fallen/1.%20Fallen%20Summary.md)
- - Added [2. The Base Element](fallen/2.%20The%20Base%20Element.md)
+ - Added [Fallen Summary](fallen/Fallen%20Summary.md)
+ - Added [The Base Element](fallen/The%20Base%20Element.md)
 
 # 0.5
 #### style.js changes